Renaming files after previous branch merge, and updating name spaces and references in other files.
