+++ SAP_oM/Excel/OpeningPsiValues.cs
@@ -26,43 +26,20 @@ using System.ComponentModel;
 using System.Linq;
 using System.Xml.Serialization;
 using BH.oM.Base;
+using BH.oM.Geometry;
 
-namespace BH.oM.Environment.SAP
+namespace BH.oM.Environment.SAP.Excel
 {
-    [Description("The details of roofs from the users excel input.")]
-    public class OpeningExcel : BHoMObject
+    [Description("PsiValues for thermal bridge.")]
+    public class OpeningPsiValues : BHoMObject
     {
-        [Description("The type of the opening.")]
+        [Description("Opening Type.")]
         public virtual string OpeningType { get; set; } = string.Empty;
 
-        [Description("The uvalue of the opening.")]
-        public virtual double UValue { get; set; } = 0;
+        [Description("Psi value forWindow lintels - tops of the windows.")]
+        public virtual List<PsiValues> PsiValues { get; set; } = new List<PsiValues>();
 
-        [Description("The gvalue of the opening.")]
-        public virtual double GValue { get; set; } = 0;
-
-        [Description("The type of glazing.")]
-        public virtual TypeOfGlazing TypeOfGlazing { get; set; } = TypeOfGlazing.DoubleLowEHard02;
-
-        [Description("The glazing gap.")]
-        public virtual GlazingGap GlazingGap { get; set; } = GlazingGap.SixteenOrMore;
-
-        [Description("The frame factor of the opening.")]
-        public virtual double FrameFactor { get; set; } = 0;
-
-        [Description("The type of frame.")]
-        public virtual TypeOfFrame FrameType { get; set; } = TypeOfFrame.Wood;
-
-        [Description("If the opening is argon filled.")]
-        public virtual bool ArgonFilled { get; set; } = false;
-
-        [Description("If the opening is krypton filled.")]
-        public virtual bool KryptonFilled { get; set; } = false;
-
-        [Description("The data source for the opening.")]
-        public virtual OpeningDataSource DataSource { get; set; } = OpeningDataSource.ManufacturerDeclaration;
-
-        [Description("Does the opening intersect with the floor?")]
+        [Description("Psi value forWindow jambs - sides of the windows (left and right).")]
         public virtual bool FloorIntersection { get; set; } = false;
     }
 }
